+++ Tests/Mocks/CardViewControllerMockDelegate.swift
@@ -2,11 +2,14 @@ import XCTest
 @testable import FramesIos
 
 class CardViewControllerMockDelegate: CardViewControllerDelegate {
-    var calledTimes = 0
-    var lastCalledWith: CkoCardTokenRequest?
+    func onSubmit(controller: CardViewController) {}
 
-    func onTapDone(controller: CardViewController, card: CkoCardTokenRequest) {
+    var calledTimes = 0
+    var lastCalledWith: CkoCardTokenResponse?
+
+    func onTapDone(controller: CardViewController, cardToken: CkoCardTokenResponse?, status: CheckoutTokenStatus) {
         calledTimes += 1
-        lastCalledWith = card
+        lastCalledWith = cardToken
     }
+
 }