+++ scripts/push.sh
@@ -2,4 +2,5 @@
 
 source ~/.rvm/scripts/rvm
 rvm use default
-pod trunk push
+pod repo add-cdn trunk 'https://cdn.cocoapods.org/'
+pod trunk push --allow-warnings